hitlad_demo/ui/requirements.txt
flask>=2.0.0
numpy>=1.21.0
pyyaml>=5.4.1
